+++ src/event-builder/types.ts
@@ -1,0 +1,23 @@
+/**
+ * Location of an element.
+ */
+export interface Point {
+  y: number;
+  x: number;
+}
+
+/**
+ * Size of an element.
+ */
+export interface Size {
+  height: number;
+  width: number;
+}
+
+/**
+ * Range of text in a text input.
+ */
+export interface TextRange {
+  start: number;
+  end: number;
+}
