+++ src/native-state.ts
@@ -1,6 +1,6 @@
 import type { HostElement } from 'test-renderer';
 
-import type { Point } from './types';
+import type { Point } from './event-builder/types';
 
 /**
  * Simulated native state for unmanaged controls.
