+++ src/user-event/clear.ts
@@ -1,10 +1,10 @@
 import type { HostElement } from 'test-renderer';
 
+import { EventBuilder } from '../event-builder';
 import { ErrorWithStack } from '../helpers/errors';
 import { isHostTextInput } from '../helpers/host-component-names';
 import { isPointerEventEnabled } from '../helpers/pointer-events';
 import { getTextInputValue, isEditableTextInput } from '../helpers/text-input';
-import { EventBuilder } from './event-builder';
 import type { UserEventInstance } from './setup';
 import { emitTypingEvents } from './type/type';
 import { dispatchEvent, wait } from './utils';
