+++ src/user-event/utils/__tests__/dispatch-event.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Text } from 'react-native';
 
 import { render, screen } from '../../..';
-import { EventBuilder } from '../../event-builder';
+import { EventBuilder } from '../../../event-builder';
 import { dispatchEvent } from '../dispatch-event';
 
 const TOUCH_EVENT = EventBuilder.Common.touch();
