+++ src/user-event/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './content-size';
 export * from './dispatch-event';
-export * from './text-range';
 export * from './wait';
