--- src/user-event/utils/text-range.ts
@@ -1,4 +1,0 @@
-export interface TextRange {
-  start: number;
-  end: number;
-}
